Refactoring and fixes

+++ external_snapshot_manager.sh
@@ -1,44 +1,53 @@
 #!/bin/bash
 
 external_snapshot_setup () {
+	local output
+	local lines
+	local diskFolder
+	local pathsArray
+	local domxml
+	local bname
+	local folder
+	local name
+	local newPath
+
 	if [[ $# != 2 ]]; then
 		echo "Usage: ${0} setup <domain> <path>"
 		return 1
 	fi
 
-	local output=`virsh snapshot-list --domain $1`
-
+	output=`virsh snapshot-list --domain $1 2>&1`
 	if [[ $? != 0 ]]; then
-		echo "$output"
+		echo -n "$output"
 		return 1
 	fi
 
-	local lines=`echo "$output" | wc -l`
+	lines=`echo -n "$output" | wc -l`
 	if [[ $lines -gt 2 ]]; then
 		echo "Function should be used before creating any snapshots."
 		return 1
 	fi
 
-	local diskFolder=`realpath "${2}/disks"`
+	diskFolder=`realpath "${2}/disks"`
 	output=`mkdir $diskFolder`
 
 	if [[ $? != 0 ]]; then
-		echo "$output"
+		echo -n "$output"
 		return 1
 	fi
 
 	echo "Created folder ${diskFolder}"
 
 	readarray -t pathsArray < <( virsh domblklist --domain $1 --details | awk 'NR>2 && $2=="disk" {print $4}' )
-	local domxml=`virsh dumpxml --domain $1`
+	domxml=`virsh dumpxml --domain $1`
 	for path in ${pathsArray[@]}; do
-		local bname=`basename $path`
-		local folder=`dirname $path`
-		local name="${bname%.*}"
-		local newPath="${diskFolder}/${name}.base"
-		output=`mv $path $newPath`
+		bname=`basename $path`
+		folder=`dirname $path`
+		name="${bname%.*}"
+		newPath="${diskFolder}/${name}.base"
+		output=`mv $path $newPath 2>&1`
 		if [[ $? != 0 ]]; then
-			echo "$output"
+			echo -n "$output"
 			return 1
 		fi
 		echo "Moved ${path} to ${newPath}"
@@ -50,6 +59,8 @@ external_snapshot_setup () {
 }
 
 external_snapshot_create () {
+	local output 
+
 	if [[ $# != 2 ]]; then
 		echo "Usage: ${0} create <domain> <name>"
 		return 1
@@ -60,10 +71,9 @@ external_snapshot_create () {
 		return 1
 	fi
 
-	local output=`virsh snapshot-create-as --domain $1 --name $2 --disk-only`
-
+	output=`virsh snapshot-create-as --domain $1 --name $2 --disk-only 2>&1`
 	if [[ $? != 0 ]]; then
-		echo "$output"
+		echo -n "$output"
 		return 1
 	fi
 
@@ -72,6 +82,15 @@ external_snapshot_create () {
 
 external_snapshot_revert () {
 	local output
+	local pathsArray
+	local domxml
+	local bname
+	local folder
+	local ext
+	local name
+	local newPath
+	local diskInfo
+	local diskType
 
 	if [[ $# != 2 ]]; then
 		echo "Usage: ${0} revert <domain> <name>"
@@ -79,37 +98,33 @@ external_snapshot_revert () {
 	fi
 
 	if [[ $2 == "base" ]]; then
-		output=`virsh dominfo --domain $1`
-		if [[ $? != 0 ]]; then
-			echo "$output"
-			return 1
-		fi
+		output=`virsh dominfo --domain $1 2>&1`
 	else
-		output=`virsh snapshot-info --domain $1 --snapshotname $2`
-		if [[ $? != 0 ]]; then
-			echo "$output"
-			return 1
-		fi
+		output=`virsh snapshot-info --domain $1 --snapshotname $2 2>&1`
+	fi
+	if [[ $? != 0 ]]; then
+		echo -n "$output"
+		return 1
 	fi
 
 	readarray -t pathsArray < <( virsh domblklist --domain $1 --details | awk 'NR>2 && $2=="disk" {print $4}' )
-	local domxml=`virsh dumpxml --domain $1`
+	domxml=`virsh dumpxml --domain $1`
 	for path in ${pathsArray[@]}; do
-		local bname=`basename $path`
-		local folder=`dirname $path`
-		local ext="${bname#*.}"
-		local name="${bname%.*}"
+		bname=`basename $path`
+		folder=`dirname $path`
+		ext="${bname#*.}"
+		name="${bname%.*}"
 		if [[ $ext == $2 ]]; then
 			echo "Can't revert to currently active snapshot!"
 			return 1
 		else
-			local newPath="${folder}/${name}.${2}"
-			local diskInfo=`qemu-img info $newPath`
+			newPath="${folder}/${name}.${2}"
+			output=`qemu-img info $newPath 2>&1`
 			if [[ $? != 0 ]]; then
-				echo $diskInfo
+				echo -n "$output"
 				return 1
 			fi
-			local diskType=`echo "$diskInfo" | grep "^file format:" | cut -d " " -f 3`
+			diskType=`echo -n "$output" | grep "^file format:" | cut -d " " -f 3`
 			domxml=`echo -n "$domxml" | xmlstarlet ed -u "/domain/devices/disk[source/@file=\"${path}\"]/driver/@type" -v $diskType`
 			domxml=`echo -n "$domxml" | xmlstarlet ed -u "/domain/devices/disk[source/@file=\"${path}\"]/source/@file" -v $newPath`
 		fi
@@ -120,6 +135,14 @@ external_snapshot_revert () {
 }
 
 external_snapshot_delete () {
+	local output
+	local pathsArray
+	local bname
+	local folder
+	local ext
+	local name
+	local newPath
+
 	if [[ $# != 2 ]]; then
 		echo "Usage: ${0} delete <domain> <name>"
 		return 1
@@ -130,23 +153,27 @@ external_snapshot_delete () {
 		return 1
 	fi
 
-	local output=`virsh snapshot-info --domain $1 --snapshotname $2`
+	output=`virsh snapshot-info --domain $1 --snapshotname $2 2>&1`
 	if [[ $? != 0 ]]; then
-		echo "$output"
+		echo -n "$output"
 		return 1
 	fi
 
 	readarray -t pathsArray < <( virsh domblklist --domain $1 --details | awk 'NR>2 && $2=="disk" {print $4}' )
 	for path in ${pathsArray[@]}; do
-		local bname=`basename $path`
-		local folder=`dirname $path`
-		local ext="${bname#*.}"
-		local name="${bname%.*}"
-		local newPath="${folder}/${name}.${2}"
+		bname=`basename $path`
+		folder=`dirname $path`
+		ext="${bname#*.}"
+		name="${bname%.*}"
+		newPath="${folder}/${name}.${2}"
 		if [[ $ext == $2 ]]; then
 			echo "The snapshot is currently active!"
 			return 1
 		else 
+			if ! [[ -f $newPath ]]; then
+				echo "File ${newPath} does not exist!"
+				return 1
+			fi
 			rm -f $newPath
 			echo "Removed ${newPath}"
 		fi
@@ -158,30 +185,40 @@ external_snapshot_delete () {
 }
 
 external_snapshot_refresh () {
+	local output
+	local pathsArray
+	local bname
+	local folder
+	local ext
+	local name
+	local backingFile
+	local backingFileBname
+	local backingFileExt
+
 	if [[ $# != 1 ]]; then
 		echo "Usage: ${0} delete <domain>"
 		return 1
 	fi
 
-	local output=`virsh dominfo --domain $1`
+	output=`virsh dominfo --domain $1 2>&1`
 	if [[ $? != 0 ]]; then
-		echo "$output"
+		echo -n "$output"
 		return 1
 	fi
 
 	readarray -t pathsArray < <( virsh domblklist --domain $1 --details | awk 'NR>2 && $2=="disk" {print $4}' )
 	for path in ${pathsArray[@]}; do
-		local bname=`basename $path`
-		local folder=`dirname $path`
-		local ext="${bname#*.}"
-		local name="${bname%.*}"
+		bname=`basename $path`
+		folder=`dirname $path`
+		ext="${bname#*.}"
+		name="${bname%.*}"
 		if [[ $ext == "base" ]]; then
 			echo "Can't refresh \"base\" images!"
 			return 1
 		fi
-		local backingFile=`qemu-img info $path | grep "backing file:" | cut -d " " -f 3`
-		local backingFileBname=`basename $backingFile`
-		local backingFileExt="${backingFileBname#*.}"
+		backingFile=`qemu-img info $path | grep "backing file:" | cut -d " " -f 3`
+		backingFileBname=`basename $backingFile`
+		backingFileExt="${backingFileBname#*.}"
 	done
 	external_snapshot_revert $1 $backingFileExt
 	if [[ $? != 0 ]]; then
@@ -201,33 +238,49 @@ external_snapshot_refresh () {
 
 external_snapshot_commit () {
 	local output
+	local pathsArray
+	local targetsArray
+	local len
+	local path
+	local target
+	local bname
+	local folder
+	local ext 
+	local name
+	local basePath
+	local topPath
 
 	if [[ $# != 3 ]]; then
 		echo "Usage: ${0} <domain> <top_snapshot> <base_snapshot>"
 		exit 1
 	fi
 
-	local blklist=`virsh domblklist --domain $1 --details`
-	readarray -t pathsArray < <( echo -n "$blklist" | awk 'NR>2 && $2=="disk" {print $4}' )
-	readarray -t targetsArray < <( echo -n "$blklist" | awk 'NR>2 && $2=="disk" {print $3}' )
-	local len=${#pathsArray[@]}
+	output=`virsh domblklist --domain $1 --details 2>&1`
+	if [[ $? != 0 ]]; then
+		echo -n "$output"
+		return 1
+	fi
+
+	readarray -t pathsArray < <( echo -n "$output" | awk 'NR>2 && $2=="disk" {print $4}' )
+	readarray -t targetsArray < <( echo -n "$output" | awk 'NR>2 && $2=="disk" {print $3}' )
+	len=${#pathsArray[@]}
 	for ((i=0; i<$len; i++)); do
-		local path=${pathsArray[$i]}
-		local target=${targetsArray[$i]}
-		local bname=`basename $path`
-		local folder=`dirname $path`
-		local ext="${bname#*.}"
-		local name="${bname%.*}"
-		local basePath="${folder}/${name}.${3}"
-		local topPath="${folder}/${name}.${2}"
+		path=${pathsArray[$i]}
+		target=${targetsArray[$i]}
+		bname=`basename $path`
+		folder=`dirname $path`
+		ext="${bname#*.}"
+		name="${bname%.*}"
+		basePath="${folder}/${name}.${3}"
+		topPath="${folder}/${name}.${2}"
 		if [[ $ext == $2 ]]; then
 			echo "Top snapshot is currently active; pivoting to base."
-			output=`virsh blockcommit --domain $1 --path $target --base $basePath --top $topPath --active --pivot`
+			output=`virsh blockcommit --domain $1 --path $target --base $basePath --top $topPath --active --pivot 2>&1`
 		else
-			output=`virsh blockcommit --domain $1 --path $target --base $basePath --top $topPath`
+			output=`virsh blockcommit --domain $1 --path $target --base $basePath --top $topPath 2>&1`
 		fi
 		if [[ $? != 0 ]]; then
-			echo $output
+			echo -n "$output"
 			return 1
 		fi
 	done
